Fix FormConfig import to not leak into cht-form
